requirements.txt
streamlit
numpy
opencv-python-headless
ezdxf
svg.path